backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-socketio==5.11.2
pydantic-settings==2.2.1
openai==1.30.1
mss==9.0.1
Pillow>=10.4.0
websockets==12.0
aiofiles==23.2.1
sqlite-utils==3.36
python-multipart==0.0.9
